requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
bcrypt
python-jose[cryptography]
email-validator
python-multipart
filestack-python